sistemazione codice

+++ backend/project/src/main/java/SyncPlanner/project/service/BookingsService.java
@@ -14,6 +14,7 @@ import org.springframework.stereotype.Service;
 
 import java.time.Instant;
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 
 @Service
@@ -30,6 +31,18 @@ public class BookingsService {
     @Autowired
     private BookingParticipantsRepo bookingParticipantsRepo;
 
+    @Autowired
+    private DiscordBotService discordBotService;
+
+    private static final Map<String, String> SEDE_TO_DISCORD_ROLE = Map.of(
+            "verona", "1308145201352671302", // ID ruolo per Verona
+            "padova", "1308145246487445555", // ID ruolo per Padova
+            "milano", "345678901234567890", // ID ruolo per Milano
+            "como", "456789012345678901",   // ID ruolo per Como
+            "napoli", "567890123456789012", // ID ruolo per Napoli
+            "roma", "678901234567890123"    // ID ruolo per Roma
+    );
+
     public List<BookingsModel> getBookings() {
         return bookingsRepo.findAll();
     }
@@ -42,6 +55,18 @@ public class BookingsService {
         UserModel user = userOptional.get();
 
         Optional<SediModel> sedeOptional = sedeRepo.findByName(bookingRequest.getSedeName());
+
+        String discordRoleId = SEDE_TO_DISCORD_ROLE.get(bookingRequest.getSedeName().toLowerCase());
+
+        if (discordRoleId != null) {
+            String message = "Ciao <@&%s>, **%s** sarà in sede il %s".formatted(
+                    discordRoleId,
+                    bookingRequest.getUserUsername(),
+                    bookingRequest.getDate()
+            );
+            discordBotService.sendMessageToThread(message);
+        }
+
         if (sedeOptional.isEmpty()) {
             throw new RuntimeException("Sede not found");
         }
@@ -71,8 +96,22 @@ public class BookingsService {
         bookingsRepo.save(booking);
     }
 
-    public void deleteBooking(Integer id) {
-        bookingParticipantsRepo.deleteByBookingId(id);
-        bookingsRepo.deleteById(id);
+    public boolean deleteBooking(Integer id) {
+        try {
+            if (bookingsRepo.findById(id).isPresent()) {
+                String discordRoleId = SEDE_TO_DISCORD_ROLE.get(bookingsRepo.findById(id).get().getSede().getName());
+                String message = "Ciao, <@&%s>, **%s** ha cancellato la prenotazione del %s".formatted(discordRoleId, bookingsRepo.findById(id).get().getUser().getUsername(), bookingsRepo.findById(id).get().getDate());
+                discordBotService.sendMessageToThread(message);
+
+                bookingParticipantsRepo.deleteByBookingId(id);
+                bookingsRepo.deleteById(id);
+
+                return true;
+            } else {
+                throw new RuntimeException("Booking not found");
+            }
+        }catch (Exception e) {
+            throw new RuntimeException("Errore durante l'eliminazione della prenotazione con id: " + id, e);
+        }
     }
 }
